identity: adding a helper package for Managed Identities
The intention here is to make this code and convention more reusable, since
the `identity` block should be treated in a common manner across different
resources.

That is, the presence of an `identity` block should indicate that an identity
is attached - and the absence should indicate that no identity should be assigned.

This commit introduces the foundation for this - and an implementation for
System Assigned Identities - which maps to the following schema:

```
resource "azurerm_example" "example" {
  identity {
    type = "SystemAssigned"
  }
}
```

With documentation showing how to use this to map to and from the Azure SDK
types - since unfortunately the Azure SDK isn't using common base types here